cloud_functions/predict/requirements.txt
google-cloud-storage 
google-cloud-automl
google-cloud-pubsub
protobuf3-to-dict